Rename the end-dash folder for now
