+++ docs/variables.markdown
@@ -6,7 +6,7 @@
 ```coffeescript
 require("../test/helper")
 Model = require("Backbone").Model
-EndDash = require("../lib/end-dash").generateTemplate
+EndDash = require("../lib/configured_parser").generateTemplate
 ```
 The above code just imports test/helper and the backbone default model.
