+++ lib/compile.js
@@ -1,8 +1,9 @@
-var EndDash = require("./end-dash")
+var ConfiguredParser = require("./configured_parser")
   , fs = require("fs")
 
-var compile = module.exports = function(module, filename, content) {
-  var Parsed = new EndDash(content, { templateName: filename.replace(/\.js\.ed(\.erb)?$/, "") })
+var compile = module.exports = function(module, filename) {
+  var content = fs.readFileSync(filename, 'utf8')
+  var Parsed = new ConfiguredParser(content, { templateName: filename.replace(/\.js\.ed(\.erb)?$/, "") })
   module.exports = Parsed.generate()
 }
 if(require.extensions) {
