+++ lib/configured_parser.js
